add more test in the chain-addr
